float commit

+++ temp.py
@@ -1,5 +1,5 @@
 tempc = input('請輸入攝氏C，我來幫你轉成華氏F唷')
-tempc = int(tempc)
+tempc = float(tempc)
 tempf = tempc * 9 / 5 + 32
-tempf = int(tempf)
+tempf = float(tempf)
 print('轉成華氏後是',tempf,'度唷！')
